reorganize compilation units
