+++ src/dllmain.cpp
@@ -1,10 +1,10 @@
 #include "expander.h"
 #include "dllmain.h"
+#include "integration.h"
 #include "lib/string.h"
 #include "lib/terminal.h"
 #include "lib/timer.h"
 #include "lib/window.h"
-#include "lib/ui/integration.h"
 #include "struct/ExecutionContext.h"
 
 extern MqlInstanceList               g_mqlInstances;        // all MQL program instances
