+++ src/lib/thread.cpp
@@ -1,8 +1,8 @@
 #include "expander.h"
+#include "integration.h"
 #include "lib/terminal.h"
 #include "lib/thread.h"
 #include "lib/window.h"
-#include "lib/ui/integration.h"
 
 
 /**
